Bump UserData and Nonce field sizes to 1024 bytes
The attestation document specification defines the user_data and
nonce fields as containing up to 1024 bytes of data, the same
size as the cert and public_key fields. Raise the parser limits
to reflect the specification.

https://docs.aws.amazon.com/enclaves/latest/user/verify-root.html

Note that the nitro hypervisor doesn't currently enforce this
limit, so it's possible to produce attestation documents with
a valid signature with more than 1k data in these fields.